requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
tavily-python
python-dotenv
pydantic
reportlab

